+++ tests/test_async.py
@@ -5,13 +5,13 @@ from .schemas import FooSchema
 from apiflask import HTTPTokenAuth
 
 
-def skip_flask1(app):
+@pytest.fixture(autouse=True)
+def skip_async_test(app):
     if not hasattr(app, 'ensure_sync'):
         pytest.skip('This test requires Flask 2.0 or higher')
 
 
 def test_async_view(app, client):
-    skip_flask1(app)
 
     @app.get('/')
     async def index():
@@ -23,7 +23,6 @@ def test_async_view(app, client):
 
 
 def test_async_error_processor(app, client):
-    skip_flask1(app)
 
     @app.error_processor
     async def custom_error_processor(e):
@@ -35,7 +34,6 @@ def test_async_error_processor(app, client):
 
 
 def test_async_spec_processor(app, client):
-    skip_flask1(app)
 
     @app.spec_processor
     async def update_spec(spec):
@@ -49,7 +47,6 @@ def test_async_spec_processor(app, client):
 
 
 def test_auth_required_on_async_view(app, client):
-    skip_flask1(app)
     auth = HTTPTokenAuth()
 
     @app.get('/')
@@ -62,7 +59,6 @@ def test_auth_required_on_async_view(app, client):
 
 
 def test_doc_on_async_view(app, client):
-    skip_flask1(app)
 
     @app.get('/')
     @app.doc(summary='Test Root Endpoint')
@@ -75,7 +71,6 @@ def test_doc_on_async_view(app, client):
 
 
 def test_input_on_async_view(app, client):
-    skip_flask1(app)
 
     @app.post('/')
     @app.input(FooSchema)
@@ -89,7 +84,6 @@ def test_input_on_async_view(app, client):
 
 
 def test_output_on_async_view(app, client):
-    skip_flask1(app)
 
     @app.get('/foo')
     @app.output(FooSchema)
@@ -110,7 +104,6 @@ def test_output_on_async_view(app, client):
 
 
 def test_async_doc_input_and_output_decorator(app, client):
-    skip_flask1(app)
 
     @app.post('/')
     @app.doc(summary='Test Root Endpoint')